requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai
sqlalchemy
flask-bcrypt
google.genai
